+++ earth_extractor/providers/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 from .base import Provider
 from .copernicus import CopernicusDataSpace, copernicus_dataspace
+from .alaskan_satellite_facility import asf
 from .nasa import nasa_cmr
 from .swisstopo import swiss_topo