+++ earth_extractor/satellites/sentinel.py
@@ -1,6 +1,6 @@
 from earth_extractor.satellites.base import Satellite
 from earth_extractor.satellites import enums
-from earth_extractor.providers import copernicus_dataspace
+from earth_extractor.providers import copernicus_dataspace, asf
 from earth_extractor import core
 import logging
 
@@ -14,7 +14,7 @@ logger.setLevel(core.config.constants.LOGLEVEL_MODULE_DEFAULT)
 
 sentinel_1: Satellite = Satellite(
     query_provider=copernicus_dataspace,
-    download_provider=copernicus_dataspace,
+    download_provider=asf,
     name=enums.Satellite.SENTINEL1,
     description="Sentinel 1",
     processing_levels=[enums.ProcessingLevel.L1, enums.ProcessingLevel.L2],